#1935 Moved SelectedRoundsProcessor into the package for test utilities
